# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for meme generation platform",
        "title": "Memology API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/memes": {
            "get": {
                "description": "Get paginated list of all memes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memes"
                ],
                "summary": "Get all memes",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Meme"
                            }
                        }
                    }
                }
            }
        },
        "/memes/generate": {
            "post": {
                "description": "Generate meme from user input using neural network. Style is optional. Returns meme with pending status and task_id for checking progress.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memes"
                ],
                "summary": "Generate new meme",
                "parameters": [
                    {
                        "description": "Meme generation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateMemeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Meme"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memes/my": {
            "get": {
                "description": "Get list of memes created by current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memes"
                ],
                "summary": "Get user memes",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Meme"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memes/styles": {
            "get": {
                "description": "Get list of available meme generation styles from neural network",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memes"
                ],
                "summary": "Get available meme styles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memes/{id}": {
            "get": {
                "description": "Get meme details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memes"
                ],
                "summary": "Get meme by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meme ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Meme"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete meme by ID (only owner can delete)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memes"
                ],
                "summary": "Delete meme",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meme ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memes/{id}/status": {
            "get": {
                "description": "Check if meme generation is completed and fetch result if ready",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memes"
                ],
                "summary": "Check meme generation status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meme ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Meme"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Meme": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "generation_time_ms": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "metrics": {
                    "$ref": "#/definitions/models.MemeMetrics"
                },
                "prompt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "style": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.MemeMetrics": {
            "type": "object",
            "properties": {
                "click_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "download_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "meme_id": {
                    "type": "string"
                },
                "other_interactions": {
                    "type": "integer"
                },
                "rating_score": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "services.CreateMemeRequest": {
            "type": "object",
            "required": [
                "prompt"
            ],
            "properties": {
                "prompt": {
                    "type": "string",
                    "example": "я купил компьютер за 1000000"
                },
                "style": {
                    "type": "string",
                    "example": "anime"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}